# Join microbenchmark